Tier prediction visualizer class for Jupyter Notebook
New class: TierVisualizer: creates widgets that plot user-supplied functions for predictions using the estimated parameters in each tier.
DataCollection now works based on dictionaries and dataframes. It now has functionality to get the DataSource class of an individual or group of a given datatype.
Renamed DataSourceBase to IndDataSourceBase.
DataSource classes now have a get_data(id) method that returns the data for an individual or group along with aux data.